Fix MC-8187
Mixin code provided nearly as is from https://github.com/Modflower/8187

The test suite was made to strictly test the saplings.
Note that the test suite only tests growing on NW as
it isn't necessary to test the other 3 corners, despite
being easy to fit in by the architecture of the code.